Moved to a library so that it can be used from other packages
