+++ SimTracker/SiPixelDigitizer/BuildFile.xml
@@ -1,0 +1,6 @@
+<use   name="FWCore/Framework"/>
+<use   name="FWCore/ParameterSet"/>
+<use   name="SimDataFormats/TrackingHit"/>
+<export>
+  <lib   name="1"/>
+</export>
