+++ SimTracker/SiPixelDigitizer/plugins/BuildFile.xml
@@ -14,6 +14,7 @@
 <use   name="CalibTracker/SiPixelESProducers"/>
 <use   name="gsl"/>
 <use   name="clhep"/>
+<use name="SimTracker/SiPixelDigitizer"/>
 <library   file="*.cc" name="SimTrackerSiPixelDigitizerPlugins">
   <flags   EDM_PLUGIN="1"/>
 </library>
